Add `PFixed` and its friends
